Add a few more Geometry tests, and move tests namespaces
